move loss back into claim
