--- loss.d.ts
@@ -1,19 +1,0 @@
-
-export type LossType = "Hurricane" |
-  "Fire" |
-  "Flood" |
-  "Water" |
-  "Wind" |
-  "Lightning" |
-  "Hail" |
-  "LiabilityInjuryPhysicalDamage" |
-  "Sinkhole" |
-  "Vandalism" |
-  "Theft" |
-  "Other" |
-  "LossAssessment" |
-  "LiabilityAnimal" |
-  "LiabilityLead" |
-  "CausedByAnimal" |
-  "PersonalInjuryProtection" |
-  "MedicalPayments"
